Suggestion Only! Page.py to include total results
Often, based on the total results, I will revise the search terms, but I can't know that easily without paging out to see how many results there were. This fixes that. (It might be useful to hide the addition when there's only a single pageful, but for a first PR, this is enough.